% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmpiricalCalibration.R
\name{doEmpiricalCalibration}
\alias{doEmpiricalCalibration}
\title{Perform empirical calibration}
\usage{
doEmpiricalCalibration(outputFolder)
}
\arguments{
\item{outputFolder}{The path to the output folder containing the results.}
}
\description{
Perform empirical calibration
}
\details{
Performs empricical calibration using the negative control outcomes, and computes the calibrated
p-values
}

