% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Main.R
\name{execute}
\alias{execute}
\title{Execute OHDSI Keppra and the Risk of Angioedema study}
\usage{
execute(connectionDetails, cdmDatabaseSchema,
  workDatabaseSchema = cdmDatabaseSchema,
  studyCohortTable = "ohdsi_alendronate_raloxifene",
  oracleTempSchema = workDatabaseSchema, outputFolder, createCohorts = TRUE,
  runAnalyses = TRUE, packageResults = TRUE, maxCores = 4)
}
\arguments{
\item{connectionDetails}{An object of type \code{connectionDetails} as created using the
\code{\link[DatabaseConnector]{createConnectionDetails}} function in the
DatabaseConnector package.}

\item{cdmDatabaseSchema}{Schema name where your patient-level data in OMOP CDM format resides.
Note that for SQL Server, this should include both the database and
schema name, for example 'cdm_data.dbo'.}

\item{workDatabaseSchema}{Schema name where intermediate data can be stored. You will need to have
write priviliges in this schema. Note that for SQL Server, this should
include both the database and schema name, for example 'cdm_data.dbo'.}

\item{studyCohortTable}{The name of the table that will be created in the work database schema.
This table will hold the exposure and outcome cohorts used in this
study.}

\item{oracleTempSchema}{Should be used in Oracle to specify a schema where the user has write
priviliges for storing temporary tables.}

\item{outputFolder}{Name of local folder to place results; make sure to use forward slashes
(/). Do not use a folder on a network drive since this greatly impacts
performance.}

\item{createCohorts}{Create the studyCohortTable table with the exposure and outcome cohorts?}

\item{runAnalyses}{Perform the cohort method analyses?}

\item{packageResults}{Package the results for sharing?}

\item{maxCores}{How many parallel cores should be used? If more cores are made available
this can speed up the analyses.}
}
\description{
Execute OHDSI Keppra and the Risk of Angioedema study
}
\details{
This function executes the OHDSI Keppra and the Risk of Angioedema study.
}
\examples{
\dontrun{
connectionDetails <- createConnectionDetails(dbms = "postgresql",
                                             user = "joe",
                                             password = "secret",
                                             server = "myserver")

execute(connectionDetails,
        cdmDatabaseSchema = "cdm_data",
        workDatabaseSchema = "results",
        studyCohortTable = "ohdsi_alendronate_raloxifene",
        oracleTempSchema = NULL,
        outputFolder = "c:/temp/study_results",
        maxCores = 4)
}

}
