% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropensityScores.R
\name{fitAllPsModels}
\alias{fitAllPsModels}
\title{Fit all propensity models}
\usage{
fitAllPsModels(workFolder, maxCores = 4)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Fit all propensity models
}
\details{
This function fits all propensity models, using the cohortMethodData objects constructed using the
\code{\link{generateAllCohortMethodDataObjects}} function.
}

