% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PropensityScores.R
\name{plotAllPsDistributions}
\alias{plotAllPsDistributions}
\title{Plot all propensity score distributions}
\usage{
plotAllPsDistributions(workFolder)
}
\arguments{
\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Plot all propensity score distributions
}
\details{
This function plots all propensity score distributions, using the propensity score objects constructed using the
\code{\link{fitAllPsModels}} function.
}

