% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CustomCmDataObjectBuilding.R
\name{constructCohortMethodDataObject}
\alias{constructCohortMethodDataObject}
\title{Construct a cohortMethodData object}
\usage{
constructCohortMethodDataObject(treatmentId, comparatorId, workFolder)
}
\arguments{
\item{treatmentId}{The cohort definition ID of the treatment cohort.}

\item{comparatorId}{The cohort definition ID of the comparator cohort.}

\item{workFolder}{Name of local folder to place results; make sure to use forward slashes
(/)}
}
\description{
Construct a cohortMethodData object
}
\details{
This function constructs a cohortMethodData object for a specific target and comparator, using the data
fetched earlier using the \code{\link{fetchAllDataFromServer}} function.
}

